🛠️ Rename ui into atoms
